@@ -15,6 +15,7 @@
 package netconf
 
 import (
+	"context"
 	"fmt"
 	"strings"
 
@@ -60,7 +61,7 @@ func (tc *TransformationContext) Close() []*sdcpb.Update {
 	return result
 }
 
-func (tc *TransformationContext) String() (result string, err error) {
+func (tc *TransformationContext) String(ctx context.Context) (result string, err error) {
 	result = "TransformationContext\n"
 	for k, v := range tc.leafLists {
 		vals := []string{}
