@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"math"
 	"strings"
 	"time"
 
@@ -12,9 +13,9 @@ import (
 	"github.com/sdcio/data-server/pkg/tree"
 	treeproto "github.com/sdcio/data-server/pkg/tree/importer/proto"
 	treetypes "github.com/sdcio/data-server/pkg/tree/types"
+	logf "github.com/sdcio/logger"
 	sdcpb "github.com/sdcio/sdc-protos/sdcpb"
 	"github.com/sdcio/sdc-protos/tree_persist"
-	log "github.com/sirupsen/logrus"
 )
 
 var (
@@ -62,6 +63,8 @@ func (d *Datastore) SdcpbTransactionIntentToInternalTI(ctx context.Context, req 
 // replaceIntent takes a Transaction and treats it as a replaceIntent, replacing the whole device configuration with the content of the given intent.
 // returns the warnings as a []string and potential errors that happend during validation / from SBI Set()
 func (d *Datastore) replaceIntent(ctx context.Context, transaction *types.Transaction) ([]string, error) {
+	log := logf.FromContext(ctx).WithValues("transaction-type", "replace")
+	ctx = logf.IntoContext(ctx, log)
 
 	// create a new TreeContext
 	tc := tree.NewTreeContext(d.schemaClient, d.Name())
@@ -95,14 +98,14 @@ func (d *Datastore) replaceIntent(ctx context.Context, transaction *types.Transa
 		return nil, err
 	}
 
-	log.Debugf("Transaction Replace: %s - finish tree insertion phase", transaction.GetTransactionId())
+	log.V(logf.VDebug).Info("transaction finish tree insertion phase")
 	err = root.FinishInsertionPhase(ctx)
 	if err != nil {
 		return nil, err
 	}
 
 	// log the tree in trace level, making it a func call to spare overhead in lower log levels.
-	log.TraceFn(func() []interface{} { return []interface{}{root.String()} })
+	log.V(logf.VTrace).Info("populated tree", "tree", root.String())
 
 	// perform validation
 	validationResult, validationStats := root.Validate(ctx, d.config.Validation, d.taskPool)
@@ -112,8 +115,7 @@ func (d *Datastore) replaceIntent(ctx context.Context, transaction *types.Transa
 	}
 
 	warnings := validationResult.WarningsStr()
-	log.Debugf("Transaction: %s - validation stats - %s", transaction.GetTransactionId(), validationStats.String())
-	log.Infof("Transaction: %s - validation passed", transaction.GetTransactionId())
+	log.Info("transaction validation passed", "transaction-id", transaction.GetTransactionId(), "stats", validationStats.String())
 
 	// we use the TargetSourceReplace, that adjustes the tree results in a way
 	// that the whole config tree is getting replaced.
@@ -127,7 +129,7 @@ func (d *Datastore) replaceIntent(ctx context.Context, transaction *types.Transa
 	// collect warnings
 	warnings = append(warnings, dataResp.GetWarnings()...)
 
-	log.Infof("ds=%s transaction=%s applied", d.Name(), transaction.GetTransactionId()+" - replace")
+	log.Info("transaction applied")
 
 	// retrieve the data that is meant to be send southbound (towards the device)
 	updates := root.GetHighestPrecedence(true)
@@ -143,6 +145,7 @@ func (d *Datastore) replaceIntent(ctx context.Context, transaction *types.Transa
 }
 
 func (d *Datastore) LoadAllButRunningIntents(ctx context.Context, root *tree.RootEntry, excludeDeviations bool) ([]string, error) {
+	log := logf.FromContext(ctx)
 
 	intentNames := []string{}
 	IntentChan := make(chan *tree_persist.Intent)
@@ -172,9 +175,9 @@ func (d *Datastore) LoadAllButRunningIntents(ctx context.Context, root *tree.Roo
 				continue
 			}
 			intentNames = append(intentNames, intent.GetIntentName())
-			log.Debugf("%s: adding intent %s to tree", d.Name(), intent.GetIntentName())
+			log.V(logf.VDebug).Info("adding intent to tree", "intent", intent.GetIntentName())
 			protoLoader := treeproto.NewProtoTreeImporter(intent)
-			log.Tracef("%s", intent.String())
+			log.V(logf.VTrace).Info("adding intent to tree", "intent", intent.String())
 			err := root.ImportConfig(ctx, nil, protoLoader, intent.GetIntentName(), intent.GetPriority(), treetypes.NewUpdateInsertFlags())
 			if err != nil {
 				return nil, err
@@ -188,6 +191,7 @@ func (d *Datastore) LoadAllButRunningIntents(ctx context.Context, root *tree.Roo
 
 // lowlevelTransactionSet
 func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *types.Transaction, dryRun bool) (*sdcpb.TransactionSetResponse, error) {
+	log := logf.FromContext(ctx)
 	// create a new TreeRoot
 	d.syncTreeMutex.Lock()
 	root, err := d.syncTree.DeepCopy(ctx)
@@ -212,7 +216,7 @@ func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *typ
 		lvs := tree.LeafVariantSlice{}
 		lvs = root.GetByOwner(intent.GetName(), lvs)
 
-		oldIntentContent := lvs.ToUpdateSlice()
+		oldIntentContent := lvs.ToPathAndUpdateSlice()
 
 		deleteVisitorPool := d.taskPool.NewVirtualPool(pool.VirtualFailFast, 1)
 		ownerDeleteMarker := tree.NewOwnerDeleteMarker(tree.NewOwnerDeleteMarkerTaskConfig(intent.GetName(), intent.GetOnlyIntended()))
@@ -231,8 +235,13 @@ func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *typ
 		// clear the owners existing explicit delete entries, retrieving the old entries for storing in the transaction for possible rollback
 		oldExplicitDeletes := root.RemoveExplicitDeletes(intent.GetName())
 
+		priority := int32(math.MaxInt32)
+		if len(oldIntentContent) > 0 {
+			priority = oldIntentContent[0].GetUpdate().Priority()
+		}
+
 		// store the old intent content in the transaction as the old intent.
-		err = transaction.AddIntentContent(intent.GetName(), types.TransactionIntentOld, oldIntentContent.GetFirstPriorityValue(), oldIntentContent, oldExplicitDeletes)
+		err = transaction.AddIntentContent(intent.GetName(), types.TransactionIntentOld, priority, oldIntentContent, oldExplicitDeletes)
 		if err != nil {
 			return nil, err
 		}
@@ -252,21 +261,21 @@ func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *typ
 	les := tree.LeafVariantSlice{}
 	les = root.GetByOwner(tree.RunningIntentName, les)
 
-	transaction.GetOldRunning().AddUpdates(les.ToUpdateSlice())
+	transaction.GetOldRunning().AddUpdates(les.ToPathAndUpdateSlice())
 
-	log.Debugf("Transaction: %s - finish tree insertion phase", transaction.GetTransactionId())
+	log.V(logf.VDebug).Info("transaction finish tree insertion phase")
 	// FinishInsertion Phase
 	err = root.FinishInsertionPhase(ctx)
 	if err != nil {
 		return nil, err
 	}
 
-	log.Debug(root.String())
+	log.V(logf.VTrace).Info("populated tree", "tree", root.String())
 
 	// perform validation
 	validationResult, validationStats := root.Validate(ctx, d.config.Validation, d.taskPool)
 
-	log.Debugf("Transaction: %s - Validation Stats: %s", transaction.GetTransactionId(), validationStats.String())
+	log.V(logf.VDebug).Info("transaction validation stats", "transaction-id", transaction.GetTransactionId(), "stats", validationStats.String())
 
 	// prepare the response struct
 	result := &sdcpb.TransactionSetResponse{
@@ -308,11 +317,11 @@ func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *typ
 		return result, ErrValidationError
 	}
 
-	log.Infof("Transaction: %s - validation passed", transaction.GetTransactionId())
+	log.Info("transaction validation passed")
 
 	// if it is a dry run, return now, skipping updating the device or the cache
 	if dryRun {
-		log.Infof("Transaction: %s - dryrun finished successfull", transaction.GetTransactionId())
+		log.Info("transaction dry-run was successful")
 		return result, nil
 	}
 
@@ -323,7 +332,7 @@ func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *typ
 	}
 	result.Warnings = append(result.Warnings, dataResp.GetWarnings()...)
 
-	log.Infof("ds=%s transaction=%s applied", d.Name(), transaction.GetTransactionId())
+	log.Info("transaction applied")
 
 	/////////////////////////////////////
 	// update intent in intended store //
@@ -331,8 +340,8 @@ func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *typ
 
 	// logging
 	updStrSl := treetypes.Map(updates.ToUpdateSlice(), func(u *treetypes.Update) string { return u.String() })
-	log.Debugf("Updates:\n%s", strings.Join(updStrSl, "\n"))
-	log.Debugf("Deletes:\n%s", strings.Join(deletes.SdcpbPaths().ToXPathSlice(), "\n"))
+	log.V(logf.VTrace).Info("generated updates", "updates", strings.Join(updStrSl, "\n"))
+	log.V(logf.VTrace).Info("generated deletes", "deletes", strings.Join(deletes.SdcpbPaths().ToXPathSlice(), "\n"))
 
 	for _, intent := range transaction.GetNewIntents() {
 		// retrieve the data that is meant to be send towards the cache
@@ -341,17 +350,17 @@ func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *typ
 
 		// logging
 		strSl := treetypes.Map(updatesOwner, func(u *treetypes.Update) string { return u.String() })
-		log.Debugf("Updates Owner: %s\n%s", intent.GetName(), strings.Join(strSl, "\n"))
+		log.V(logf.VTrace).Info("updates owner", "updates-owner", strSl, "\n")
 
 		delSl := deletesOwner.ToXPathSlice()
-		log.Debugf("Deletes Owner: %s\n%s", intent.GetName(), strings.Join(delSl, "\n"))
+		log.V(logf.VTrace).Info("deletes owner", "deletes-owner", delSl, "\n")
 
 		protoIntent, err := root.TreeExport(intent.GetName(), intent.GetPriority(), intent.Deviation())
 		switch {
 		case errors.Is(err, tree.ErrorIntentNotPresent):
 			err = d.cacheClient.IntentDelete(ctx, intent.GetName(), intent.GetDeleteIgnoreNonExisting())
 			if err != nil {
-				log.Warnf("failed deleting intent from store for %s: %v", d.Name(), err)
+				log.Error(err, "failed deleting intent from store")
 			}
 			continue
 		case err != nil:
@@ -369,11 +378,11 @@ func (d *Datastore) lowlevelTransactionSet(ctx context.Context, transaction *typ
 		return nil, err
 	}
 
-	log.Infof("ds=%s transaction=%s: completed", d.Name(), transaction.GetTransactionId())
+	log.Info("transaction completed")
 	// start the rollback ticker only if it was not already a rollback transaction.
 	if !transaction.IsRollback() {
 		// start the RollbackTimer
-		err = transaction.StartRollbackTimer()
+		err = transaction.StartRollbackTimer(ctx)
 		if err != nil {
 			return nil, err
 		}
@@ -406,7 +415,7 @@ func (d *Datastore) writeBackSyncTree(ctx context.Context, updates tree.LeafVari
 
 	// export the synctree
 	newRunningIntent, err := d.syncTree.TreeExport(tree.RunningIntentName, tree.RunningValuesPrio, false)
-	if err != nil {
+	if err != nil && err != tree.ErrorIntentNotPresent {
 		return err
 	}
 	// write the synctree to disk
@@ -420,21 +429,22 @@ func (d *Datastore) writeBackSyncTree(ctx context.Context, updates tree.LeafVari
 }
 
 func (d *Datastore) TransactionSet(ctx context.Context, transactionId string, transactionIntents []*types.TransactionIntent, replaceIntent *types.TransactionIntent, transactionTimeout time.Duration, dryRun bool) (*sdcpb.TransactionSetResponse, error) {
+	log := logf.FromContext(ctx)
 	var err error
 	var transaction *types.Transaction
 	var transactionGuard *types.TransactionGuard
 
-	log.Infof("Transaction: %s - start", transactionId)
+	log.Info("transaction start")
 
 	// create a new Transaction with the given transaction id
 	transaction = types.NewTransaction(transactionId, d.transactionManager)
 	// set the timeout on the transaction
-	transaction.SetTimeout(transactionTimeout)
+	transaction.SetTimeout(ctx, transactionTimeout)
 
 	if !dryRun {
 		// try locking the datastore if it is locked return the specific ErrDatastoreLocked error.
 		if !d.dmutex.TryLock() {
-			log.Infof("Transaction: %s - abort (%v)", transactionId, ErrDatastoreLocked)
+			log.Error(ErrDatastoreLocked, "transaction abort")
 			return nil, ErrDatastoreLocked
 		}
 		defer d.dmutex.Unlock()
@@ -445,7 +455,7 @@ func (d *Datastore) TransactionSet(ctx context.Context, transactionId string, tr
 			select {
 			case <-ctx.Done():
 				// Context was canceled or timed out
-				log.Errorf("Transaction: %s - context canceled or timed out: %v", transactionId, ctx.Err())
+				log.Error(ctx.Err(), "transaction context canceled or timed out", ctx.Err())
 				return nil, ErrContextDone
 			default:
 				// Start a transaction and prepare to cancel it if any error occurs
@@ -469,7 +479,7 @@ func (d *Datastore) TransactionSet(ctx context.Context, transactionId string, tr
 	if transaction.GetReplace() != nil {
 		replaceWarn, err := d.replaceIntent(ctx, transaction)
 		if err != nil {
-			log.Errorf("error setting replace intent: %v", err)
+			log.Error(err, "error setting replace intent")
 			return nil, err
 		}
 		// TODO: do something with these warnings
@@ -478,7 +488,7 @@ func (d *Datastore) TransactionSet(ctx context.Context, transactionId string, tr
 
 	err = transaction.AddTransactionIntents(transactionIntents, types.TransactionIntentNew)
 	if err != nil {
-		log.Errorf("error adding intents to transaction: %v", err)
+		log.Error(err, "error adding intents to transaction")
 		return nil, err
 	}
 
@@ -495,12 +505,12 @@ func (d *Datastore) TransactionSet(ctx context.Context, transactionId string, tr
 	// if it is a validation error, we need to send the response while not successing the transaction guard
 	// since validation errors are transported in the response itself, not in the seperate error
 	if errors.Is(err, ErrValidationError) {
-		log.Errorf("Transaction: %s - validation failed\n%s", transactionId, strings.Join(response.GetErrors(), "\n"))
+		log.Error(fmt.Errorf("%s", strings.Join(response.GetErrors(), ", ")), "transaction validation failed")
 		return response, nil
 	}
 	// if it is any other error, return a regular error
 	if err != nil {
-		log.Errorf("error executing transaction: %v", err)
+		log.Error(err, "error executing transaction")
 		return nil, err
 	}
 
@@ -509,7 +519,7 @@ func (d *Datastore) TransactionSet(ctx context.Context, transactionId string, tr
 		// Mark the transaction as successfully committed
 		transactionGuard.Success()
 
-		log.Infof("Transaction: %s - transacted", transactionId)
+		log.Info("transaction success")
 	}
 	return response, err
 }
@@ -529,7 +539,8 @@ func updateToSdcpbUpdate(lvs tree.LeafVariantSlice) ([]*sdcpb.Update, error) {
 }
 
 func (d *Datastore) TransactionConfirm(ctx context.Context, transactionId string) error {
-	log.Infof("Transaction %s - Confirm", transactionId)
+	log := logf.FromContext(ctx)
+	log.Info("transaction confirm")
 
 	if !d.dmutex.TryLock() {
 		return ErrDatastoreLocked
@@ -540,7 +551,8 @@ func (d *Datastore) TransactionConfirm(ctx context.Context, transactionId string
 }
 
 func (d *Datastore) TransactionCancel(ctx context.Context, transactionId string) error {
-	log.Infof("Transaction %s - Cancel", transactionId)
+	log := logf.FromContext(ctx)
+	log.Info("transaction cancel")
 
 	if !d.dmutex.TryLock() {
 		return ErrDatastoreLocked
