@@ -21,35 +21,39 @@ import (
 	"time"
 
 	cconfig "github.com/sdcio/cache/pkg/config"
-	log "github.com/sirupsen/logrus"
-
 	"github.com/sdcio/data-server/pkg/cache"
+	logf "github.com/sdcio/logger"
 )
 
 func (s *Server) createCacheClient(ctx context.Context) {
+	log := logf.FromContext(ctx)
+	log = log.WithValues("cache-type", s.config.Cache.Type)
+	ctx = logf.IntoContext(ctx, log)
 START:
 	var err error
 	switch s.config.Cache.Type {
 	default:
+		log.Error(nil, "invalid cache type")
 		fmt.Fprintf(os.Stderr, "unknown cache type: %s", s.config.Cache.Type)
 		os.Exit(1)
 	case "local":
 		err = s.createLocalCacheClient(ctx)
 		if err != nil {
-			log.Errorf("failed to initialize a local cache client: %v", err)
+			log.Error(err, "failed to initialize a local cache client")
 			time.Sleep(time.Second)
 			goto START
 		}
-		log.Infof("local cache created")
 	}
 }
 
-func (s *Server) createLocalCacheClient(_ context.Context) error {
+func (s *Server) createLocalCacheClient(ctx context.Context) error {
+	log := logf.FromContext(ctx)
 	var err error
-	log.Infof("initializing local cache client")
+	log.Info("initializing local cache client", "store-type", s.config.Cache.StoreType, "directory", s.config.Cache.Dir)
 	s.cacheClient, err = cache.NewLocalCache(&cconfig.CacheConfig{
 		StoreType: s.config.Cache.StoreType,
 		Dir:       s.config.Cache.Dir,
 	})
+	log.Info("local cache created")
 	return err
 }
