@@ -21,20 +21,23 @@ import (
 	"strings"
 	"time"
 
-	sdcpb "github.com/sdcio/sdc-protos/sdcpb"
-	log "github.com/sirupsen/logrus"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/peer"
-	"google.golang.org/grpc/status"
-
 	"github.com/sdcio/data-server/pkg/config"
 	"github.com/sdcio/data-server/pkg/datastore"
 	targettypes "github.com/sdcio/data-server/pkg/datastore/target/types"
+	"github.com/sdcio/data-server/pkg/utils"
+	logf "github.com/sdcio/logger"
+	sdcpb "github.com/sdcio/sdc-protos/sdcpb"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/peer"
+	"google.golang.org/grpc/status"
 )
 
 // datastore
 func (s *Server) ListDataStore(ctx context.Context, req *sdcpb.ListDataStoreRequest) (*sdcpb.ListDataStoreResponse, error) {
-	log.Debug("Received ListDataStoreRequest")
+	log := logf.FromContext(ctx).WithName("ListDataStore")
+	ctx = logf.IntoContext(ctx, log)
+
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
 
 	datastores := s.datastores.GetDatastoreAll()
 	rs := make([]*sdcpb.GetDataStoreResponse, 0, len(datastores))
@@ -51,7 +54,13 @@ func (s *Server) ListDataStore(ctx context.Context, req *sdcpb.ListDataStoreRequ
 }
 
 func (s *Server) GetDataStore(ctx context.Context, req *sdcpb.GetDataStoreRequest) (*sdcpb.GetDataStoreResponse, error) {
-	log.Debugf("Received GetDataStoreRequest: %v", req)
+	log := logf.FromContext(ctx).WithName("GetDataStore")
+	log = log.WithValues(
+		"datastore-name", req.GetDatastoreName(),
+	)
+	ctx = logf.IntoContext(ctx, log)
+
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
 	name := req.GetDatastoreName()
 	if name == "" {
 		return nil, status.Error(codes.InvalidArgument, "missing datastore name attribute")
@@ -64,7 +73,18 @@ func (s *Server) GetDataStore(ctx context.Context, req *sdcpb.GetDataStoreReques
 }
 
 func (s *Server) CreateDataStore(ctx context.Context, req *sdcpb.CreateDataStoreRequest) (*sdcpb.CreateDataStoreResponse, error) {
-	log.Debugf("Received CreateDataStoreRequest: %v", req)
+	log := logf.FromContext(ctx).WithName("CreateDataStore")
+	log = log.WithValues(
+		"datastore-name", req.GetDatastoreName(),
+	)
+	ctx = logf.IntoContext(ctx, log)
+
+	log.Info("creating datastore",
+		"datastore-schema", req.GetSchema(),
+		"datastore-target", req.GetTarget(),
+	)
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
+
 	name := req.GetDatastoreName()
 	lName := len(name)
 	if lName == 0 {
@@ -183,12 +203,19 @@ func (s *Server) CreateDataStore(ctx context.Context, req *sdcpb.CreateDataStore
 	if err != nil {
 		return nil, err
 	}
-	s.datastores.AddDatastore(ds)
+	err = s.datastores.AddDatastore(ds)
+	if err != nil {
+		return nil, err
+	}
 	return &sdcpb.CreateDataStoreResponse{}, nil
 }
 
 func (s *Server) DeleteDataStore(ctx context.Context, req *sdcpb.DeleteDataStoreRequest) (*sdcpb.DeleteDataStoreResponse, error) {
-	log.Debugf("Received DeleteDataStoreRequest: %v", req)
+	log := logf.FromContext(ctx).WithName("DeleteDataStore")
+	ctx = logf.IntoContext(ctx, log)
+
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
+
 	name := req.GetName()
 	if name == "" {
 		return nil, status.Error(codes.InvalidArgument, "missing datastore name attribute")
@@ -201,10 +228,14 @@ func (s *Server) DeleteDataStore(ctx context.Context, req *sdcpb.DeleteDataStore
 
 	err = ds.Stop(s.ctx)
 	if err != nil {
-		log.Errorf("failed to stop datastore %s: %v", name, err)
+		log.Error(err, "failed to stop datastore")
 	}
-	s.datastores.DeleteDatastore(ctx, name)
-	log.Infof("deleted datastore %s", name)
+	err = s.datastores.DeleteDatastore(ctx, name)
+	if err != nil {
+		log.Error(err, "failed to delete datastore")
+		return nil, fmt.Errorf("failed to delete datastore: %w", err)
+	}
+	log.Info("deleted datastore")
 
 	return &sdcpb.DeleteDataStoreResponse{}, nil
 }
@@ -212,7 +243,10 @@ func (s *Server) DeleteDataStore(ctx context.Context, req *sdcpb.DeleteDataStore
 func (s *Server) WatchDeviations(req *sdcpb.WatchDeviationRequest, stream sdcpb.DataServer_WatchDeviationsServer) error {
 	ctx := stream.Context()
 	p, ok := peer.FromContext(ctx)
-	log.Debugf("Received WatchDeviationRequest from peer %s: %v", p.Addr.String(), req)
+	log := logf.FromContext(ctx).WithName("WatchDeviations").WithValues("peer", p.String())
+	ctx = logf.IntoContext(ctx, log)
+
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
 
 	peerInfo, ok := peer.FromContext(ctx)
 	if !ok {
@@ -222,15 +256,19 @@ func (s *Server) WatchDeviations(req *sdcpb.WatchDeviationRequest, stream sdcpb.
 		return status.Errorf(codes.InvalidArgument, "missing datastore name")
 	}
 
+	if len(req.GetName()) == 0 {
+		return status.Errorf(codes.InvalidArgument, "missing datastore name")
+	}
+
 	ds, err := s.datastores.GetDataStore(req.GetName()[0])
 	if err != nil {
-		log.Error(err)
+		log.Error(err, "failed to get datastore")
 		return status.Errorf(codes.NotFound, "unknown datastore")
 	}
 
 	err = ds.WatchDeviations(req, stream)
 	if err != nil {
-		log.Error(err)
+		log.Error(err, "failed to watch deviations")
 	}
 	<-stream.Context().Done()
 	ds.StopDeviationsWatch(peerInfo.Addr.String())
@@ -265,6 +303,10 @@ func (s *Server) datastoreToRsp(ctx context.Context, ds *datastore.Datastore) (*
 }
 
 func (s *Server) BlameConfig(ctx context.Context, req *sdcpb.BlameConfigRequest) (*sdcpb.BlameConfigResponse, error) {
+	log := logf.FromContext(ctx).WithName("BlameConfig")
+	ctx = logf.IntoContext(ctx, log)
+
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
 
 	if req.GetDatastoreName() == "" {
 		return nil, status.Errorf(codes.InvalidArgument, "missing datastore name")
