@@ -8,8 +8,9 @@ import (
 	"github.com/sdcio/data-server/pkg/datastore"
 	"github.com/sdcio/data-server/pkg/datastore/types"
 	"github.com/sdcio/data-server/pkg/tree"
+	"github.com/sdcio/data-server/pkg/utils"
+	logf "github.com/sdcio/logger"
 	sdcpb "github.com/sdcio/sdc-protos/sdcpb"
-	log "github.com/sirupsen/logrus"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/peer"
 	"google.golang.org/grpc/status"
@@ -17,7 +18,21 @@ import (
 
 func (s *Server) TransactionSet(ctx context.Context, req *sdcpb.TransactionSetRequest) (*sdcpb.TransactionSetResponse, error) {
 	pr, _ := peer.FromContext(ctx)
-	log.Debugf("received TransactionSetRequest request %v from peer %s", req, pr.Addr.String())
+
+	log := logf.FromContext(ctx).WithName("TransactionSet")
+	log = log.WithName("transaction").WithValues(
+		"transaction-id", req.GetTransactionId(),
+	)
+	ctx = logf.IntoContext(ctx, log)
+
+	log.Info("TransactionSet",
+		"transaction-datastore-name", req.GetDatastoreName(),
+		"transaction-dry-run", req.GetDryRun(),
+		"transaction-timeout", req.GetTimeout(),
+		"transaction-peer", pr.String(),
+	)
+
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
 
 	if req.GetDatastoreName() == "" {
 		return nil, status.Error(codes.InvalidArgument, "missing datastore name")
@@ -70,7 +85,19 @@ func (s *Server) TransactionSet(ctx context.Context, req *sdcpb.TransactionSetRe
 
 func (s *Server) TransactionConfirm(ctx context.Context, req *sdcpb.TransactionConfirmRequest) (*sdcpb.TransactionConfirmResponse, error) {
 	pr, _ := peer.FromContext(ctx)
-	log.Debugf("received TransactionConfirm request %v from peer %s", req, pr.Addr.String())
+
+	log := logf.FromContext(ctx).WithName("TransactionConfirm")
+	log = log.WithName("transaction").WithValues(
+		"transaction-id", req.GetTransactionId(),
+	)
+	ctx = logf.IntoContext(ctx, log)
+
+	log.Info("TransactionConfirm",
+		"transaction-datastore-name", req.GetDatastoreName(),
+		"transaction-peer", pr.String(),
+	)
+
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
 
 	if req.GetDatastoreName() == "" {
 		return nil, status.Error(codes.InvalidArgument, "missing datastore name")
@@ -87,7 +114,19 @@ func (s *Server) TransactionConfirm(ctx context.Context, req *sdcpb.TransactionC
 
 func (s *Server) TransactionCancel(ctx context.Context, req *sdcpb.TransactionCancelRequest) (*sdcpb.TransactionCancelResponse, error) {
 	pr, _ := peer.FromContext(ctx)
-	log.Debugf("received TransactionCancel request %v from peer %s", req, pr.Addr.String())
+
+	log := logf.FromContext(ctx).WithName("TransactionCancel")
+	log = log.WithName("transaction").WithValues(
+		"transaction-id", req.GetTransactionId(),
+	)
+	ctx = logf.IntoContext(ctx, log)
+
+	log.Info("TransactionCancel",
+		"transaction-datastore-name", req.GetDatastoreName(),
+		"transaction-peer", pr.String(),
+	)
+
+	log.V(logf.VDebug).Info("received request", "raw-request", utils.FormatProtoJSON(req))
 
 	if req.GetDatastoreName() == "" {
 		return nil, status.Error(codes.InvalidArgument, "missing datastore name")
