@@ -62,9 +62,9 @@ func Test_Entry(t *testing.T) {
 		IsRootBased: false,
 	}
 
-	u1 := types.NewUpdate(p1, desc, int32(100), "me", int64(9999999))
-	u2 := types.NewUpdate(p2, desc, int32(99), "me", int64(444))
-	u3 := types.NewUpdate(p2, desc, int32(98), "me", int64(88))
+	u1 := types.NewUpdate(nil, desc, int32(100), "me", int64(9999999))
+	u2 := types.NewUpdate(nil, desc, int32(99), "me", int64(444))
+	u3 := types.NewUpdate(nil, desc, int32(98), "me", int64(88))
 
 	tc := NewTreeContext(scb, "foo")
 
@@ -73,8 +73,8 @@ func Test_Entry(t *testing.T) {
 		t.Error(err)
 	}
 
-	for _, u := range []*types.Update{u1, u2, u3} {
-		_, err = root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+	for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1, u1), types.NewPathAndUpdate(p2, u2), types.NewPathAndUpdate(p2, u3)} {
+		_, err = root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 		if err != nil {
 			t.Error(err)
 		}
@@ -165,14 +165,14 @@ func Test_Entry_One(t *testing.T) {
 		IsRootBased: true,
 	}
 
-	u0o1 := types.NewUpdate(p0o1, testhelper.GetStringTvProto("ethernet-1/1"), prio100, owner1, ts1)
-	u0o2 := types.NewUpdate(p0o2, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner2, ts1)
-	u1 := types.NewUpdate(p1, desc1, prio100, owner1, ts1)
-	u1_1 := types.NewUpdate(p1_1, testhelper.GetUIntTvProto(9), prio100, owner1, ts1)
-	u2 := types.NewUpdate(p2, desc2, prio100, owner1, ts1)
-	u2_1 := types.NewUpdate(p2_1, testhelper.GetUIntTvProto(10), prio100, owner1, ts1)
-	u3 := types.NewUpdate(p2, desc3, prio50, owner2, ts1)
-	u3_1 := types.NewUpdate(p2_1, testhelper.GetUIntTvProto(10), prio50, owner2, ts1)
+	u0o1 := types.NewUpdate(nil, testhelper.GetStringTvProto("ethernet-1/1"), prio100, owner1, ts1)
+	u0o2 := types.NewUpdate(nil, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner2, ts1)
+	u1 := types.NewUpdate(nil, desc1, prio100, owner1, ts1)
+	u1_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(9), prio100, owner1, ts1)
+	u2 := types.NewUpdate(nil, desc2, prio100, owner1, ts1)
+	u2_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(10), prio100, owner1, ts1)
+	u3 := types.NewUpdate(nil, desc3, prio50, owner2, ts1)
+	u3_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(10), prio50, owner2, ts1)
 
 	tc := NewTreeContext(scb, "foo")
 
@@ -182,8 +182,16 @@ func Test_Entry_One(t *testing.T) {
 	}
 
 	// start test
-	for _, u := range []*types.Update{u0o1, u1, u1_1, u2, u2_1, u3, u3_1} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+	for _, u := range []*types.PathAndUpdate{
+		types.NewPathAndUpdate(p0o1, u0o1),
+		types.NewPathAndUpdate(p1, u1),
+		types.NewPathAndUpdate(p1_1, u1_1),
+		types.NewPathAndUpdate(p2, u2),
+		types.NewPathAndUpdate(p2_1, u2_1),
+		types.NewPathAndUpdate(p2, u3),
+		types.NewPathAndUpdate(p2_1, u3_1),
+	} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 		if err != nil {
 			t.Error(err)
 		}
@@ -202,7 +210,14 @@ func Test_Entry_One(t *testing.T) {
 		o1Le = root.GetByOwner(owner1, o1Le)
 		o1 := LeafEntriesToUpdates(o1Le)
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{u0o1, u2, u2_1, u1, u1_1}, o1); diff != "" {
+		// if diff := testhelper.DiffUpdates([]*types.Update{u0o1, u2, u2_1, u1, u1_1}, o1); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(p0o1, u0o1),
+			types.NewPathAndUpdate(p2, u2),
+			types.NewPathAndUpdate(p2_1, u2_1),
+			types.NewPathAndUpdate(p1, u1),
+			types.NewPathAndUpdate(p1_1, u1_1),
+		}, o1); diff != "" {
 			t.Errorf("root.GetByOwner(owner1) mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -212,7 +227,10 @@ func Test_Entry_One(t *testing.T) {
 		o2Le = root.GetByOwner(owner2, o2Le)
 		o2 := LeafEntriesToUpdates(o2Le)
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{u0o2, u3_1, u3}, o2); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(p0o2, u0o2),
+			types.NewPathAndUpdate(p2_1, u3_1),
+			types.NewPathAndUpdate(p2, u3)}, o2); diff != "" {
 			t.Errorf("root.GetByOwner(owner2) mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -221,7 +239,14 @@ func Test_Entry_One(t *testing.T) {
 
 		highprec := root.GetHighestPrecedence(true)
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{u0o2, u1, u1_1, u3, u3_1}, highprec.ToUpdateSlice()); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(p0o2, u0o2),
+			types.NewPathAndUpdate(p1, u1),
+			types.NewPathAndUpdate(p1_1, u1_1),
+			types.NewPathAndUpdate(p2, u3),
+			types.NewPathAndUpdate(p2_1, u3_1),
+		},
+			highprec.ToUpdateSlice()); diff != "" {
 			t.Errorf("root.GetHighesPrio() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -271,9 +296,9 @@ func Test_Entry_Two(t *testing.T) {
 		IsRootBased: true,
 	}
 
-	u0 := types.NewUpdate(p0, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
-	u1 := types.NewUpdate(p1, desc3, prio50, owner1, ts1)
-	u1_1 := types.NewUpdate(p1_1, testhelper.GetUIntTvProto(10), prio50, owner1, ts1)
+	u0 := types.NewUpdate(nil, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
+	u1 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u1_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(10), prio50, owner1, ts1)
 
 	tc := NewTreeContext(scb, "foo")
 
@@ -283,15 +308,15 @@ func Test_Entry_Two(t *testing.T) {
 	}
 
 	// start test add "existing" data
-	for _, u := range []*types.Update{u1} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsExisting)
+	for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1, u1)} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsExisting)
 		if err != nil {
 			t.Error(err)
 		}
 	}
 
 	// add incomming set intent reques data
-	overwriteDesc := testhelper.GetStringTvProto("Owerwrite Description")
+	overwriteDesc := testhelper.GetStringTvProto("Overwrite Description")
 
 	// adding a new Update with same owner and priority with different value
 	pn1 := &sdcpb.Path{
@@ -300,12 +325,12 @@ func Test_Entry_Two(t *testing.T) {
 			sdcpb.NewPathElem("subinterface", map[string]string{"index": "10"}),
 			sdcpb.NewPathElem("description", nil),
 		},
-		IsRootBased: false,
+		IsRootBased: true,
 	}
-	n1 := types.NewUpdate(pn1, overwriteDesc, prio50, owner1, ts1)
+	n1 := types.NewUpdate(nil, overwriteDesc, prio50, owner1, ts1)
 
-	for _, u := range []*types.Update{n1} {
-		_, err = root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+	for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(pn1, n1)} {
+		_, err = root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 		if err != nil {
 			t.Error(err)
 		}
@@ -321,7 +346,11 @@ func Test_Entry_Two(t *testing.T) {
 	highprec := root.GetHighestPrecedence(true)
 
 	// diff the result with the expected
-	if diff := testhelper.DiffUpdates([]*types.Update{n1, u0, u1_1}, highprec.ToUpdateSlice()); diff != "" {
+	if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+		types.NewPathAndUpdate(pn1, n1),
+		types.NewPathAndUpdate(p0, u0),
+		types.NewPathAndUpdate(p1_1, u1_1),
+	}, highprec.ToUpdateSlice()); diff != "" {
 		t.Errorf("root.GetHighesPrio() mismatch (-want +got):\n%s", diff)
 	}
 }
@@ -418,15 +447,15 @@ func Test_Entry_Three(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u0 := types.NewUpdate(p0, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
-	u1 := types.NewUpdate(p1, desc3, prio50, owner1, ts1)
-	u1_1 := types.NewUpdate(p1_1, testhelper.GetUIntTvProto(10), prio50, owner1, ts1)
-	u2 := types.NewUpdate(p2, desc3, prio50, owner1, ts1)
-	u2_1 := types.NewUpdate(p2_1, testhelper.GetUIntTvProto(11), prio50, owner1, ts1)
-	u3 := types.NewUpdate(p3, desc3, prio50, owner1, ts1)
-	u3_1 := types.NewUpdate(p3_1, testhelper.GetUIntTvProto(12), prio50, owner1, ts1)
-	u4 := types.NewUpdate(p4, desc3, prio50, owner1, ts1)
-	u4_1 := types.NewUpdate(p4_1, testhelper.GetUIntTvProto(13), prio50, owner1, ts1)
+	u0 := types.NewUpdate(nil, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
+	u1 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u1_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(10), prio50, owner1, ts1)
+	u2 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u2_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(11), prio50, owner1, ts1)
+	u3 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u3_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(12), prio50, owner1, ts1)
+	u4 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u4_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(13), prio50, owner1, ts1)
 
 	p1r := &sdcpb.Path{
 		Elem: []*sdcpb.PathElem{
@@ -461,10 +490,10 @@ func Test_Entry_Three(t *testing.T) {
 		IsRootBased: true,
 	}
 
-	u1r := types.NewUpdate(p1r, desc3, RunningValuesPrio, RunningIntentName, ts1)
-	u2r := types.NewUpdate(p2r, desc3, RunningValuesPrio, RunningIntentName, ts1)
-	u3r := types.NewUpdate(p3r, desc3, RunningValuesPrio, RunningIntentName, ts1)
-	u4r := types.NewUpdate(p4r, desc3, RunningValuesPrio, RunningIntentName, ts1)
+	u1r := types.NewUpdate(nil, desc3, RunningValuesPrio, RunningIntentName, ts1)
+	u2r := types.NewUpdate(nil, desc3, RunningValuesPrio, RunningIntentName, ts1)
+	u3r := types.NewUpdate(nil, desc3, RunningValuesPrio, RunningIntentName, ts1)
+	u4r := types.NewUpdate(nil, desc3, RunningValuesPrio, RunningIntentName, ts1)
 
 	tc := NewTreeContext(scb, "foo")
 
@@ -474,16 +503,21 @@ func Test_Entry_Three(t *testing.T) {
 	}
 
 	// start test add "existing" data
-	for _, u := range []*types.Update{u1, u2, u3, u4} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsExisting)
+	for _, u := range []*types.PathAndUpdate{
+		types.NewPathAndUpdate(p1, u1),
+		types.NewPathAndUpdate(p2, u2),
+		types.NewPathAndUpdate(p3, u3),
+		types.NewPathAndUpdate(p4, u4)} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsExisting)
 		if err != nil {
 			t.Error(err)
 		}
 	}
 
 	// start test add "existing" data as running
-	for _, u := range []*types.Update{u1r, u2r, u3r, u4r} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsExisting)
+	for _, u := range []*types.PathAndUpdate{
+		types.NewPathAndUpdate(p1r, u1r), types.NewPathAndUpdate(p2r, u2r), types.NewPathAndUpdate(p3r, u3r), types.NewPathAndUpdate(p4r, u4r)} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsExisting)
 		if err != nil {
 			t.Error(err)
 		}
@@ -502,7 +536,17 @@ func Test_Entry_Three(t *testing.T) {
 		highpri := root.GetHighestPrecedence(false)
 
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{u0, u1, u1_1, u2, u2_1, u3, u3_1, u4, u4_1}, highpri.ToUpdateSlice()); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(p0, u0),
+			types.NewPathAndUpdate(p1, u1),
+			types.NewPathAndUpdate(p1_1, u1_1),
+			types.NewPathAndUpdate(p2, u2),
+			types.NewPathAndUpdate(p2_1, u2_1),
+			types.NewPathAndUpdate(p3, u3),
+			types.NewPathAndUpdate(p3_1, u3_1),
+			types.NewPathAndUpdate(p4, u4),
+			types.NewPathAndUpdate(p4_1, u4_1),
+		}, highpri.ToUpdateSlice()); diff != "" {
 			t.Errorf("root.GetHighesPrio() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -515,7 +559,7 @@ func Test_Entry_Three(t *testing.T) {
 		highpri := root.GetHighestPrecedence(true)
 
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{}, highpri.ToUpdateSlice()); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{}, highpri.ToUpdateSlice()); diff != "" {
 			t.Errorf("root.GetHighesPrio() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -542,7 +586,7 @@ func Test_Entry_Three(t *testing.T) {
 			sdcpb.NewPathElem("subinterface", map[string]string{"index": "10"}),
 			sdcpb.NewPathElem("description", nil),
 		},
-		IsRootBased: false,
+		IsRootBased: true,
 	}
 
 	pn2 := &sdcpb.Path{
@@ -551,14 +595,14 @@ func Test_Entry_Three(t *testing.T) {
 			sdcpb.NewPathElem("subinterface", map[string]string{"index": "11"}),
 			sdcpb.NewPathElem("description", nil),
 		},
-		IsRootBased: false,
+		IsRootBased: true,
 	}
 
-	n1 := types.NewUpdate(pn1, overwriteDesc, prio50, owner1, ts1)
-	n2 := types.NewUpdate(pn2, overwriteDesc, prio50, owner1, ts1)
+	n1 := types.NewUpdate(nil, overwriteDesc, prio50, owner1, ts1)
+	n2 := types.NewUpdate(nil, overwriteDesc, prio50, owner1, ts1)
 
-	for _, u := range []*types.Update{n1, n2} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+	for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(pn1, n1), types.NewPathAndUpdate(pn2, n2)} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 		if err != nil {
 			t.Error(err)
 		}
@@ -582,7 +626,14 @@ func Test_Entry_Three(t *testing.T) {
 		highPri := LeafEntriesToUpdates(highPriLe)
 
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{n1, n2, u0, u1_1, u2_1}, highPri); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(pn1, n1),
+			types.NewPathAndUpdate(pn2, n2),
+			types.NewPathAndUpdate(p0, u0),
+			types.NewPathAndUpdate(p1_1, u1_1),
+			types.NewPathAndUpdate(p2_1, u2_1),
+		},
+			highPri); diff != "" {
 			t.Errorf("root.GetByOwner() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -590,7 +641,7 @@ func Test_Entry_Three(t *testing.T) {
 	t.Run("Check the old entries are gone", func(t *testing.T) {
 		highpri := root.GetHighestPrecedence(true)
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{n1, n2}, highpri.ToUpdateSlice()); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{types.NewPathAndUpdate(pn1, n1), types.NewPathAndUpdate(pn2, n2)}, highpri.ToUpdateSlice()); diff != "" {
 			t.Errorf("root.GetHighesPrio() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -714,20 +765,20 @@ func Test_Entry_Four(t *testing.T) {
 		IsRootBased: true,
 	}
 
-	u1o1_0 := types.NewUpdate(p1o1_0, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
-	u1o1 := types.NewUpdate(p1o1, desc3, prio50, owner1, ts1)
-	u1o1_1 := types.NewUpdate(p1o1_1, testhelper.GetUIntTvProto(10), prio50, owner1, ts1)
-	u2o1 := types.NewUpdate(p2o1, desc3, prio50, owner1, ts1)
-	u2o1_1 := types.NewUpdate(p2o1_1, testhelper.GetUIntTvProto(11), prio50, owner1, ts1)
-	u3 := types.NewUpdate(p3, desc3, prio50, owner1, ts1)
-	u3_1 := types.NewUpdate(p3_1, testhelper.GetUIntTvProto(12), prio50, owner1, ts1)
-	u4 := types.NewUpdate(p4, desc3, prio50, owner1, ts1)
-	u4_1 := types.NewUpdate(p4_1, testhelper.GetUIntTvProto(13), prio50, owner1, ts1)
+	u1o1_0 := types.NewUpdate(nil, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
+	u1o1 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u1o1_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(10), prio50, owner1, ts1)
+	u2o1 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u2o1_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(11), prio50, owner1, ts1)
+	u3 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u3_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(12), prio50, owner1, ts1)
+	u4 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
+	u4_1 := types.NewUpdate(nil, testhelper.GetUIntTvProto(13), prio50, owner1, ts1)
 
 	// u1o2_0 := types.NewUpdate(p1o2_0, testhelper.GetStringTvProto("ethernet-1/1"), prio55, owner2, ts1)
-	u1o2 := types.NewUpdate(p1o2, desc3, prio55, owner2, ts1)
+	u1o2 := types.NewUpdate(nil, desc3, prio55, owner2, ts1)
 	// u1o2_1 := types.NewUpdate(p1o2_1, testhelper.GetUIntTvProto(10), prio55, owner2, ts1)
-	u2o2 := types.NewUpdate(p2o2, desc3, prio55, owner2, ts1)
+	u2o2 := types.NewUpdate(nil, desc3, prio55, owner2, ts1)
 	// p2o2_1, err := scb.ToPath(ctx, []string{"interface", "ethernet-1/1", "subinterface", "11", "index"})
 	if err != nil {
 		t.Fatal(err)
@@ -742,8 +793,19 @@ func Test_Entry_Four(t *testing.T) {
 	}
 
 	// start test add "existing" data
-	for _, u := range []*types.Update{u1o1, u2o1, u1o1_1, u2o1_1, u3, u4, u3_1, u4_1, u1o2, u2o2} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsExisting)
+	for _, u := range []*types.PathAndUpdate{
+		types.NewPathAndUpdate(p1o1, u1o1),
+		types.NewPathAndUpdate(p2o1, u2o1),
+		types.NewPathAndUpdate(p1o1_1, u1o1_1),
+		types.NewPathAndUpdate(p2o1_1, u2o1_1),
+		types.NewPathAndUpdate(p3, u3),
+		types.NewPathAndUpdate(p4, u4),
+		types.NewPathAndUpdate(p3_1, u3_1),
+		types.NewPathAndUpdate(p4_1, u4_1),
+		types.NewPathAndUpdate(p1o2, u1o2),
+		types.NewPathAndUpdate(p2o2, u2o2),
+	} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsExisting)
 		if err != nil {
 			t.Error(err)
 		}
@@ -762,7 +824,17 @@ func Test_Entry_Four(t *testing.T) {
 		highprec := root.GetHighestPrecedence(false)
 
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{u1o1_0, u1o1, u2o1, u3, u4, u1o1_1, u2o1_1, u3_1, u4_1}, highprec.ToUpdateSlice()); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(p1o1_0, u1o1_0),
+			types.NewPathAndUpdate(p1o1, u1o1),
+			types.NewPathAndUpdate(p2o1, u2o1),
+			types.NewPathAndUpdate(p3, u3),
+			types.NewPathAndUpdate(p4, u4),
+			types.NewPathAndUpdate(p1o1_1, u1o1_1),
+			types.NewPathAndUpdate(p2o1_1, u2o1_1),
+			types.NewPathAndUpdate(p3_1, u3_1),
+			types.NewPathAndUpdate(p4_1, u4_1),
+		}, highprec.ToUpdateSlice()); diff != "" {
 			t.Errorf("root.GetHighestPrecedence() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -793,7 +865,7 @@ func Test_Entry_Four(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	n1 := types.NewUpdate(pn1, overwriteDesc, prio50, owner1, ts1)
+	n1 := types.NewUpdate(nil, overwriteDesc, prio50, owner1, ts1)
 	// n2_1 := types.NewUpdate([]string{"interface", "ethernet-1/1", "subinterface", "11", "index"}, testhelper.GetIntTvProto(t, 11), prio50, owner1, ts1)
 	pn2 := &sdcpb.Path{
 		Elem: []*sdcpb.PathElem{
@@ -803,10 +875,10 @@ func Test_Entry_Four(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	n2 := types.NewUpdate(pn2, overwriteDesc, prio50, owner1, ts1)
+	n2 := types.NewUpdate(nil, overwriteDesc, prio50, owner1, ts1)
 
-	for _, u := range []*types.Update{n1, n2} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+	for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(pn1, n1), types.NewPathAndUpdate(pn2, n2)} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 		if err != nil {
 			t.Error(err)
 		}
@@ -829,7 +901,9 @@ func Test_Entry_Four(t *testing.T) {
 		highPri := LeafEntriesToUpdates(highPriLe)
 
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{n1, n2, u1o1_0, u1o1_1, u2o1_1}, highPri); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(pn1, n1), types.NewPathAndUpdate(pn2, n2), types.NewPathAndUpdate(p1o1_0, u1o1_0), types.NewPathAndUpdate(p1o1_1, u1o1_1), types.NewPathAndUpdate(p2o1_1, u2o1_1),
+		}, highPri); diff != "" {
 			t.Errorf("root.GetByOwner() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -837,7 +911,17 @@ func Test_Entry_Four(t *testing.T) {
 	t.Run("Check the old entries are gone from highest", func(t *testing.T) {
 		highpri := root.GetHighestPrecedence(false)
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{u1o1_0, n1, n2, u2o1_1, u1o1_1, u3, u3_1, u4, u4_1}, highpri.ToUpdateSlice()); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(p1o1_0, u1o1_0),
+			types.NewPathAndUpdate(pn1, n1),
+			types.NewPathAndUpdate(pn2, n2),
+			types.NewPathAndUpdate(p2o1_1, u2o1_1),
+			types.NewPathAndUpdate(p1o1_1, u1o1_1),
+			types.NewPathAndUpdate(p3, u3),
+			types.NewPathAndUpdate(p3_1, u3_1),
+			types.NewPathAndUpdate(p4, u4),
+			types.NewPathAndUpdate(p4_1, u4_1),
+		}, highpri.ToUpdateSlice()); diff != "" {
 			t.Errorf("root.GetHighestPrecedence() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -845,7 +929,13 @@ func Test_Entry_Four(t *testing.T) {
 	t.Run("Check the old entries are gone from highest (Only New Or Updated)", func(t *testing.T) {
 		highpri := root.GetHighestPrecedence(true)
 		// diff the result with the expected
-		if diff := testhelper.DiffUpdates([]*types.Update{u1o1_0, n1, n2, u2o1_1, u1o1_1}, highpri.ToUpdateSlice()); diff != "" {
+		if diff := testhelper.DiffUpdates([]*types.PathAndUpdate{
+			types.NewPathAndUpdate(p1o1_0, u1o1_0),
+			types.NewPathAndUpdate(pn1, n1),
+			types.NewPathAndUpdate(pn2, n2),
+			types.NewPathAndUpdate(p2o1_1, u2o1_1),
+			types.NewPathAndUpdate(p1o1_1, u1o1_1),
+		}, highpri.ToUpdateSlice()); diff != "" {
 			t.Errorf("root.GetHighestPrecedence() mismatch (-want +got):\n%s", diff)
 		}
 	})
@@ -890,11 +980,11 @@ func Test_Validation_Leaflist_Min_Max(t *testing.T) {
 				IsRootBased: true,
 			}
 
-			u1 := types.NewUpdate(p1, leaflistval, prio50, owner1, ts1)
+			u1 := types.NewUpdate(nil, leaflistval, prio50, owner1, ts1)
 
 			// start test add "existing" data
-			for _, u := range []*types.Update{u1} {
-				_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsExisting)
+			for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1, u1)} {
+				_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsExisting)
 				if err != nil {
 					t.Fatal(err)
 				}
@@ -946,11 +1036,11 @@ func Test_Validation_Leaflist_Min_Max(t *testing.T) {
 				IsRootBased: true,
 			}
 
-			u1 := types.NewUpdate(p1, leaflistval, prio50, owner1, ts1)
+			u1 := types.NewUpdate(nil, leaflistval, prio50, owner1, ts1)
 
 			// start test add "existing" data
-			for _, u := range []*types.Update{u1} {
-				_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsExisting)
+			for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1, u1)} {
+				_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsExisting)
 				if err != nil {
 					t.Fatal(err)
 				}
@@ -1001,11 +1091,11 @@ func Test_Validation_Leaflist_Min_Max(t *testing.T) {
 				IsRootBased: true,
 			}
 
-			u1 := types.NewUpdate(p1, leaflistval, prio50, owner1, ts1)
+			u1 := types.NewUpdate(nil, leaflistval, prio50, owner1, ts1)
 
 			// start test add "existing" data
-			for _, u := range []*types.Update{u1} {
-				_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsExisting)
+			for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1, u1)} {
+				_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsExisting)
 				if err != nil {
 					t.Fatal(err)
 				}
@@ -1046,7 +1136,7 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u1 := types.NewUpdate(p1, desc3, prio50, owner1, ts1)
+	u1 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
 
 	p2 := &sdcpb.Path{
 		Elem: []*sdcpb.PathElem{
@@ -1055,7 +1145,7 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u2 := types.NewUpdate(p2, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
+	u2 := types.NewUpdate(nil, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
 
 	p3 := &sdcpb.Path{
 		Elem: []*sdcpb.PathElem{
@@ -1065,7 +1155,7 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u3 := types.NewUpdate(p3, testhelper.GetUIntTvProto(0), prio50, owner1, ts1)
+	u3 := types.NewUpdate(nil, testhelper.GetUIntTvProto(0), prio50, owner1, ts1)
 
 	p4 := &sdcpb.Path{
 		Elem: []*sdcpb.PathElem{
@@ -1075,7 +1165,7 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u4 := types.NewUpdate(p4, desc3, prio50, owner1, ts1)
+	u4 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
 
 	p5 := &sdcpb.Path{
 		Elem: []*sdcpb.PathElem{
@@ -1085,7 +1175,7 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u5 := types.NewUpdate(p5, testhelper.GetUIntTvProto(1), prio50, owner1, ts1)
+	u5 := types.NewUpdate(nil, testhelper.GetUIntTvProto(1), prio50, owner1, ts1)
 
 	p6 := &sdcpb.Path{
 		Elem: []*sdcpb.PathElem{
@@ -1095,7 +1185,7 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u6 := types.NewUpdate(p6, desc3, prio50, owner1, ts1)
+	u6 := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
 
 	tc := NewTreeContext(scb, "foo")
 
@@ -1105,14 +1195,22 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 	}
 
 	// start test add "existing" data
-	for _, u := range []*types.Update{u1, u2, u3, u4, u5, u6} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsExisting)
+	for _, u := range []*types.PathAndUpdate{
+		types.NewPathAndUpdate(p1, u1),
+		types.NewPathAndUpdate(p2, u2),
+		types.NewPathAndUpdate(p3, u3),
+		types.NewPathAndUpdate(p4, u4),
+		types.NewPathAndUpdate(p5, u5),
+		types.NewPathAndUpdate(p6, u6),
+	} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsExisting)
 		if err != nil {
 			t.Fatal(err)
 		}
 		// add also as running
-		runUpd := u.DeepCopy().SetOwner(RunningIntentName).SetPriority(RunningValuesPrio)
-		_, err = root.AddUpdateRecursive(ctx, runUpd.Path(), runUpd, flagsExisting)
+		runUpd := u.DeepCopy()
+		runUpd.GetUpdate().SetOwner(RunningIntentName).SetPriority(RunningValuesPrio)
+		_, err = root.AddUpdateRecursive(ctx, runUpd.GetPath(), runUpd.GetUpdate(), flagsExisting)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1135,7 +1233,7 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u1n := types.NewUpdate(p1n, desc3, prio50, owner1, ts1)
+	u1n := types.NewUpdate(nil, desc3, prio50, owner1, ts1)
 
 	p2n := &sdcpb.Path{
 		Elem: []*sdcpb.PathElem{
@@ -1144,11 +1242,11 @@ func Test_Entry_Delete_Aggregation(t *testing.T) {
 		},
 		IsRootBased: true,
 	}
-	u2n := types.NewUpdate(p2n, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
+	u2n := types.NewUpdate(nil, testhelper.GetStringTvProto("ethernet-1/1"), prio50, owner1, ts1)
 
 	// start test add "new" / request data
-	for _, u := range []*types.Update{u1n, u2n} {
-		_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+	for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1n, u1n), types.NewPathAndUpdate(p2n, u2n)} {
+		_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1193,9 +1291,6 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 	owner1 := "owner1"
 	owner2 := "owner2"
 	ts := int64(0)
-	path := &sdcpb.Path{
-		Elem: []*sdcpb.PathElem{sdcpb.NewPathElem("firstPathElem", nil)},
-	}
 
 	// test that if highes prio is to be deleted, that the second highes is returned,
 	// because thats an update.
@@ -1203,8 +1298,8 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 		func(t *testing.T) {
 			lv := newLeafVariants(&TreeContext{}, nil)
 
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 2, owner1, ts), flagsExisting, nil))
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 1, owner2, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 2, owner1, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 1, owner2, ts), flagsExisting, nil))
 			lv.les[1].MarkDelete(false)
 
 			le := lv.GetHighestPrecedence(true, false, false)
@@ -1221,7 +1316,7 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 	t.Run("Single entry thats also marked for deletion",
 		func(t *testing.T) {
 			lv := newLeafVariants(&TreeContext{}, nil)
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 1, owner1, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 1, owner1, ts), flagsExisting, nil))
 			lv.les[0].MarkDelete(false)
 
 			le := lv.GetHighestPrecedence(true, false, false)
@@ -1237,9 +1332,9 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 	t.Run("New Low Prio IsUpdate OnlyChanged True",
 		func(t *testing.T) {
 			lv := newLeafVariants(&TreeContext{}, nil)
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 5, owner1, ts), flagsExisting, nil))
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 6, owner2, ts), flagsNew, nil))
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, RunningValuesPrio, RunningIntentName, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 5, owner1, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 6, owner2, ts), flagsNew, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, RunningValuesPrio, RunningIntentName, ts), flagsExisting, nil))
 
 			le := lv.GetHighestPrecedence(true, false, false)
 
@@ -1254,8 +1349,8 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 	t.Run("New Low Prio IsUpdate OnlyChanged False",
 		func(t *testing.T) {
 			lv := newLeafVariants(&TreeContext{}, nil)
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 5, owner1, ts), flagsExisting, nil))
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 6, owner1, ts), flagsNew, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 5, owner1, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 6, owner1, ts), flagsNew, nil))
 
 			le := lv.GetHighestPrecedence(false, false, false)
 
@@ -1271,9 +1366,9 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 		func(t *testing.T) {
 			lv := newLeafVariants(&TreeContext{}, nil)
 
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 5, owner1, ts), flagsExisting, nil))
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 6, owner2, ts), flagsNew, nil))
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, RunningValuesPrio, RunningIntentName, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 5, owner1, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 6, owner2, ts), flagsNew, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, RunningValuesPrio, RunningIntentName, ts), flagsExisting, nil))
 
 			le := lv.GetHighestPrecedence(true, false, false)
 
@@ -1289,8 +1384,8 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 		func(t *testing.T) {
 			lv := newLeafVariants(&TreeContext{}, nil)
 
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 5, owner1, ts), flagsExisting, nil))
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 6, owner2, ts), flagsNew, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 5, owner1, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 6, owner2, ts), flagsNew, nil))
 
 			le := lv.GetHighestPrecedence(true, false, false)
 
@@ -1303,8 +1398,8 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 	t.Run("New Low Prio IsNew OnlyChanged == False",
 		func(t *testing.T) {
 			lv := newLeafVariants(&TreeContext{}, nil)
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 5, owner1, ts), flagsExisting, nil))
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 6, owner2, ts), flagsNew, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 5, owner1, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 6, owner2, ts), flagsNew, nil))
 
 			le := lv.GetHighestPrecedence(false, false, false)
 
@@ -1331,9 +1426,9 @@ func TestLeafVariants_GetHighesPrio(t *testing.T) {
 	t.Run("secondhighes populated if highes was first",
 		func(t *testing.T) {
 			lv := newLeafVariants(&TreeContext{}, nil)
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 1, owner1, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 1, owner1, ts), flagsExisting, nil))
 			lv.les[0].MarkDelete(false)
-			lv.Add(NewLeafEntry(types.NewUpdate(path, nil, 2, owner2, ts), flagsExisting, nil))
+			lv.Add(NewLeafEntry(types.NewUpdate(nil, nil, 2, owner2, ts), flagsExisting, nil))
 
 			le := lv.GetHighestPrecedence(true, false, false)
 
@@ -1496,9 +1591,13 @@ func Test_sharedEntryAttributes_SdcpbPath(t *testing.T) {
 						},
 					},
 				},
+				IsRootBased: true,
 			}
 
-			cmp.Diff(e00.SdcpbPath().ToXPath(false), cmpPath.ToXPath(false))
+			if diff := cmp.Diff(e00.SdcpbPath().ToXPath(false), cmpPath.ToXPath(false)); diff != "" {
+				t.Errorf("sharedEntryAttributes.SdcpbPath() mismatch (-want +got):\n%s", diff)
+				return
+			}
 		},
 	)
 
@@ -1516,9 +1615,13 @@ func Test_sharedEntryAttributes_SdcpbPath(t *testing.T) {
 						Name: "description",
 					},
 				},
+				IsRootBased: true,
 			}
 
-			cmp.Diff(e00desc.SdcpbPath().ToXPath(false), cmpPath.ToXPath(false))
+			if diff := cmp.Diff(e00desc.SdcpbPath().ToXPath(false), cmpPath.ToXPath(false)); diff != "" {
+				t.Errorf("sharedEntryAttributes.SdcpbPath() mismatch (-want +got):\n%s", diff)
+				return
+			}
 		},
 	)
 
@@ -1529,17 +1632,21 @@ func Test_sharedEntryAttributes_SdcpbPath(t *testing.T) {
 					{
 						Name: "doublekey",
 						Key: map[string]string{
-							"key1": "key1",
 							"key2": "key2",
+							"key1": "key1",
 						},
 					},
 					{
 						Name: "mandato",
 					},
 				},
+				IsRootBased: true,
 			}
 
-			cmp.Diff(dkkv.SdcpbPath().ToXPath(false), cmpPath.ToXPath(false))
+			if diff := cmp.Diff(dkkv.SdcpbPath().ToXPath(false), cmpPath.ToXPath(false)); diff != "" {
+				t.Errorf("sharedEntryAttributes.SdcpbPath() mismatch (-want +got):\n%s", diff)
+				return
+			}
 		},
 	)
 
@@ -1570,10 +1677,11 @@ func Test_Validation_String_Pattern(t *testing.T) {
 
 			leafval := testhelper.GetStringTvProto("data123")
 
-			u1 := types.NewUpdate(&sdcpb.Path{Elem: []*sdcpb.PathElem{sdcpb.NewPathElem("patterntest", nil)}}, leafval, prio50, owner1, ts1)
+			p1 := &sdcpb.Path{Elem: []*sdcpb.PathElem{sdcpb.NewPathElem("patterntest", nil)}}
+			u1 := types.NewUpdate(nil, leafval, prio50, owner1, ts1)
 
-			for _, u := range []*types.Update{u1} {
-				_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+			for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1, u1)} {
+				_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 				if err != nil {
 					t.Fatal(err)
 				}
@@ -1607,11 +1715,11 @@ func Test_Validation_String_Pattern(t *testing.T) {
 			}
 
 			leafval := testhelper.GetStringTvProto("hallo F")
+			p1 := &sdcpb.Path{Elem: []*sdcpb.PathElem{sdcpb.NewPathElem("patterntest", nil)}}
+			u1 := types.NewUpdate(nil, leafval, prio50, owner1, ts1)
 
-			u1 := types.NewUpdate(&sdcpb.Path{Elem: []*sdcpb.PathElem{sdcpb.NewPathElem("patterntest", nil)}}, leafval, prio50, owner1, ts1)
-
-			for _, u := range []*types.Update{u1} {
-				_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+			for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1, u1)} {
+				_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 				if err != nil {
 					t.Fatal(err)
 				}
@@ -1704,11 +1812,11 @@ func Test_Validation_Deref(t *testing.T) {
 			}
 
 			leafval := testhelper.GetStringTvProto("data123")
+			p1 := &sdcpb.Path{Elem: []*sdcpb.PathElem{sdcpb.NewPathElem("patterntest", nil)}}
+			u1 := types.NewUpdate(nil, leafval, prio50, owner1, ts1)
 
-			u1 := types.NewUpdate(&sdcpb.Path{Elem: []*sdcpb.PathElem{sdcpb.NewPathElem("patterntest", nil)}}, leafval, prio50, owner1, ts1)
-
-			for _, u := range []*types.Update{u1} {
-				_, err := root.AddUpdateRecursive(ctx, u.Path(), u, flagsNew)
+			for _, u := range []*types.PathAndUpdate{types.NewPathAndUpdate(p1, u1)} {
+				_, err := root.AddUpdateRecursive(ctx, u.GetPath(), u.GetUpdate(), flagsNew)
 				if err != nil {
 					t.Fatal(err)
 				}
@@ -1727,6 +1835,191 @@ func Test_Validation_Deref(t *testing.T) {
 			errs := validationResult.ErrorsStr()
 			if len(errs) != 1 {
 				t.Errorf("expected 1 error but got %d, %v", len(errs), errs)
+			}
+		},
+	)
+}
+
+// Test_Validation_MultiKey_Pattern validates that each key in a multi-key list
+// has its own pattern applied correctly, rather than all keys being validated
+// against the same pattern. This test covers the bug fix in checkAndCreateKeysAsLeafs
+// where keys were being sorted, causing the key-to-value mapping to break.
+func Test_Validation_MultiKey_Pattern(t *testing.T) {
+	prio50 := int32(50)
+	owner1 := "OwnerOne"
+	ts1 := int64(9999999)
+
+	ctx := context.TODO()
+
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	scb, err := testhelper.GetSchemaClientBound(t, mockCtrl)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	t.Run("MultiKey_Pattern_Valid",
+		func(t *testing.T) {
+			tc := NewTreeContext(scb, owner1)
+			root, err := NewTreeRoot(ctx, tc)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			// Create a static-route list entry with 3 keys with mutually exclusive patterns:
+			// - owner: must match pattern 'owner-.*'
+			// - ipv6-prefix: must match pattern 'prefix-.*'
+			// - next-hop: must match pattern 'nexthop-.*'
+			// The keys are defined in schema as: key "owner ipv6-prefix next-hop"
+
+			bfdEnabledVal := &sdcpb.TypedValue{Value: &sdcpb.TypedValue_BoolVal{BoolVal: true}}
+
+			p1 := &sdcpb.Path{Elem: []*sdcpb.PathElem{
+				sdcpb.NewPathElem("ipv6", nil),
+				sdcpb.NewPathElem("static-route", map[string]string{
+					"owner":       "owner-up",        // Should match owner pattern 'owner-.*'
+					"ipv6-prefix": "prefix-default",  // Should match ipv6-prefix pattern 'prefix-.*'
+					"next-hop":    "nexthop-gateway", // Should match next-hop pattern 'nexthop-.*'
+				}),
+				sdcpb.NewPathElem("bfd-enabled", nil),
+			}}
+			// First static route: /ipv6/static-route[owner=owner-up][ipv6-prefix=prefix-default][next-hop=nexthop-gateway]/bfd-enabled
+			u1 := types.NewUpdate(
+				nil,
+				bfdEnabledVal,
+				prio50,
+				owner1,
+				ts1,
+			)
+
+			_, err = root.AddUpdateRecursive(ctx, p1, u1, flagsNew)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			p2 := &sdcpb.Path{Elem: []*sdcpb.PathElem{
+				sdcpb.NewPathElem("ipv6", nil),
+				sdcpb.NewPathElem("static-route", map[string]string{
+					"owner":       "owner-up", // Same owner and prefix
+					"ipv6-prefix": "prefix-default",
+					"next-hop":    "nexthop-backup", // Different next-hop
+				}),
+				sdcpb.NewPathElem("bfd-enabled", nil),
+			}}
+			// Second static route: /ipv6/static-route[owner=owner-up][ipv6-prefix=prefix-default][next-hop=nexthop-backup]/bfd-enabled
+			u2 := types.NewUpdate(
+				nil,
+				bfdEnabledVal,
+				prio50,
+				owner1,
+				ts1,
+			)
+
+			_, err = root.AddUpdateRecursive(ctx, p2, u2, flagsNew)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			err = root.FinishInsertionPhase(ctx)
+			if err != nil {
+				t.Error(err)
+			}
+
+			validationResult, _ := root.Validate(context.TODO(), validationConfig)
+
+			// Should have no errors - all keys match their respective patterns
+			if validationResult.HasErrors() {
+				errs := validationResult.ErrorsStr()
+				t.Errorf("expected 0 errors but got %d: %v", len(errs), errs)
+			}
+		},
+	)
+
+	t.Run("MultiKey_Pattern_Invalid_Owner",
+		func(t *testing.T) {
+			tc := NewTreeContext(scb, owner1)
+			root, err := NewTreeRoot(ctx, tc)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			// Create a static-route list entry with invalid owner (doesn't match pattern)
+			bfdEnabledVal := &sdcpb.TypedValue{Value: &sdcpb.TypedValue_BoolVal{BoolVal: true}}
+
+			// Path: /ipv6/static-route[owner=invalid][ipv6-prefix=prefix-default][next-hop=nexthop-gateway]/bfd-enabled
+			// owner="invalid" should FAIL 'owner-.*' pattern (doesn't start with 'owner-')
+			p1 := &sdcpb.Path{Elem: []*sdcpb.PathElem{
+				sdcpb.NewPathElem("ipv6", nil),
+				sdcpb.NewPathElem("static-route", map[string]string{
+					"owner":       "invalid",         // Should FAIL 'owner-.*' pattern
+					"ipv6-prefix": "prefix-default",  // Should match 'prefix-.*' pattern
+					"next-hop":    "nexthop-gateway", // Should match 'nexthop-.*' pattern
+				}),
+				sdcpb.NewPathElem("bfd-enabled", nil),
+			}}
+			u1 := types.NewUpdate(
+				nil,
+				bfdEnabledVal,
+				prio50,
+				owner1,
+				ts1,
+			)
+
+			// Pattern validation happens during AddUpdateRecursive (in checkAndCreateKeysAsLeafs)
+			// so we expect an error here
+			_, err = root.AddUpdateRecursive(ctx, p1, u1, flagsNew)
+			if err == nil {
+				t.Fatal("expected error for owner pattern mismatch, but got none")
+			}
+
+			// Verify the error message mentions the pattern
+			if !strings.Contains(err.Error(), "does not match patterns") {
+				t.Errorf("expected pattern mismatch error, got: %v", err)
+			}
+		},
+	)
+
+	t.Run("MultiKey_Pattern_Invalid_NextHop",
+		func(t *testing.T) {
+			tc := NewTreeContext(scb, owner1)
+			root, err := NewTreeRoot(ctx, tc)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			// Create a static-route list entry with invalid next-hop (doesn't match pattern)
+			bfdEnabledVal := &sdcpb.TypedValue{Value: &sdcpb.TypedValue_BoolVal{BoolVal: true}}
+
+			p1 := &sdcpb.Path{Elem: []*sdcpb.PathElem{
+				sdcpb.NewPathElem("ipv6", nil),
+				sdcpb.NewPathElem("static-route", map[string]string{
+					"owner":       "owner-up",       // Should match 'owner-.*' pattern
+					"ipv6-prefix": "prefix-default", // Should match 'prefix-.*' pattern
+					"next-hop":    "invalid",        // Should FAIL 'nexthop-.*' pattern
+				}),
+				sdcpb.NewPathElem("bfd-enabled", nil),
+			}}
+			// Path: /ipv6/static-route[owner=owner-up][ipv6-prefix=prefix-default][next-hop=invalid]/bfd-enabled
+			// next-hop="invalid" should FAIL 'nexthop-.*' pattern (doesn't start with 'nexthop-')
+			u1 := types.NewUpdate(
+				nil,
+				bfdEnabledVal,
+				prio50,
+				owner1,
+				ts1,
+			)
+
+			// Pattern validation happens during AddUpdateRecursive (in checkAndCreateKeysAsLeafs)
+			// so we expect an error here
+			_, err = root.AddUpdateRecursive(ctx, p1, u1, flagsNew)
+			if err == nil {
+				t.Fatal("expected error for next-hop pattern mismatch, but got none")
+			}
+
+			// Verify the error message mentions the pattern
+			if !strings.Contains(err.Error(), "does not match patterns") {
+				t.Errorf("expected pattern mismatch error, got: %v", err)
 			}
 		},
 	)
