@@ -10,8 +10,8 @@ import (
 	"strconv"
 	"strings"
 
+	logf "github.com/sdcio/logger"
 	sdcpb "github.com/sdcio/sdc-protos/sdcpb"
-	log "github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -150,7 +150,7 @@ func (c *Converter) ExpandUpdate(ctx context.Context, upd *sdcpb.Update) ([]*sdc
 
 		// We expect that all identityrefs are sent by schema-server as a identityref type now, not string
 		if rsp.Field.GetType().Type == "identityref" && upd.GetValue().GetStringVal() != "" {
-			upd.Value, err = Convert(upd.GetValue().GetStringVal(), rsp.Field.GetType())
+			upd.Value, err = Convert(ctx, upd.GetValue().GetStringVal(), rsp.Field.GetType())
 			if err != nil {
 				return nil, err
 			}
@@ -209,6 +209,7 @@ func (c *Converter) ExpandUpdateKeysAsLeaf(ctx context.Context, upd *sdcpb.Updat
 }
 
 func (c *Converter) ExpandContainerValue(ctx context.Context, p *sdcpb.Path, jv any, cs *sdcpb.SchemaElem_Container) ([]*sdcpb.Update, error) {
+	log := logf.FromContext(ctx)
 	// log.Debugf("expanding jsonVal %T | %v | %v", jv, jv, p)
 	switch jv := jv.(type) {
 	case string:
@@ -397,7 +398,7 @@ func (c *Converter) ExpandContainerValue(ctx context.Context, p *sdcpb.Path, jv 
 		}
 		return upds, nil
 	default:
-		log.Warnf("unexpected json type cast %T", jv)
+		log.Error(nil, "unexpected json type cast", "type", reflect.TypeOf(jv).String())
 		return nil, nil
 	}
 }
@@ -488,18 +489,14 @@ func convertStringToTv(schemaType *sdcpb.SchemaLeafType, v string, ts uint64) (*
 			Value:     &sdcpb.TypedValue_BoolVal{BoolVal: b},
 		}, nil
 	case "decimal64":
-		arr := strings.SplitN(v, ".", 2)
-		digits, err := strconv.ParseInt(arr[0], 10, 64)
+		decimalVal, err := ParseDecimal64(v)
 		if err != nil {
 			return nil, err
 		}
-		precision64, err := strconv.ParseUint(arr[1], 10, 32)
-		if err != nil {
-			return nil, err
-		}
-		precision := uint32(precision64)
 		return &sdcpb.TypedValue{
-			Value: &sdcpb.TypedValue_DecimalVal{DecimalVal: &sdcpb.Decimal64{Digits: digits, Precision: precision}},
+			Value: &sdcpb.TypedValue_DecimalVal{
+				DecimalVal: decimalVal,
+			},
 		}, nil
 	case "identityref":
 		before, name, found := strings.Cut(v, ":")
@@ -605,6 +602,7 @@ func getLeafList(s string, cs *sdcpb.SchemaElem_Container) (*sdcpb.LeafListSchem
 }
 
 func getChild(ctx context.Context, name string, cs *sdcpb.SchemaElem_Container, scb SchemaClientBound) (any, bool) {
+	log := logf.FromContext(ctx)
 
 	searchNames := []string{name}
 	if i := strings.Index(name, ":"); i >= 0 {
@@ -616,7 +614,7 @@ func getChild(ctx context.Context, name string, cs *sdcpb.SchemaElem_Container, 
 			for _, c := range cs.Container.GetChildren() {
 				rsp, err := scb.GetSchemaSdcpbPath(ctx, &sdcpb.Path{Elem: []*sdcpb.PathElem{{Name: c}}})
 				if err != nil {
-					log.Errorf("Failed to get schema object %s: %v", c, err)
+					log.Error(err, "failed to get schema object", "schema-object", c)
 					return "", false
 				}
 				switch rsp := rsp.GetSchema().Schema.(type) {
